Finish quiz refactor
